requirements.txt
crewai>=0.11.0
crewai-tools>=0.0.20
openai>=1.12.0
pydub>=0.25.1
python-dotenv>=1.0.0
agentops>=0.1.0
google-api-python-client>=2.118.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
markdown>=3.5.2
elevenlabs>=0.2.27
langchain>=0.1.0
langchain-openai>=0.0.2
typing_extensions>=4.5.0
langchain-community==0.3.24
